requirements.txt
# Core
numpy<2
pandas>=1.5.3
scikit-learn>=1.3.0
lightgbm>=4.3.2

# MLOps
mlflow>=2.6.1

# LLM / RAG
openai>=1.0.0
faiss-cpu>=1.7.4
chromadb>=0.4.7

# Backend / API
fastapi>=0.111.2
uvicorn>=0.23.1
pydantic>=2.6.2
python-dotenv>=1.0.1

# UI
streamlit>=1.26.0

# Extras / visualization
matplotlib>=3.8.0
seaborn>=0.12.2